+++ SF3Lib/Models/Structs/X1/Battle/CharacterAttackScoreBonus.cs
@@ -17,7 +17,7 @@ namespace SF3.Models.Structs.X1 {
         [TableViewModelColumn(addressField: nameof(_valueAddr), displayOrder: 1, displayFormat: "X2")]
         [BulkCopy]
         public byte Value {
-            get => (byte) Data.GetUInt8(_valueAddr);
+            get => Data.GetUInt8(_valueAddr);
             set => Data.SetUInt8(_valueAddr, value);
         }
     }
