+++ SF3Lib/Models/Structs/X1/Battle/CharacterMoveTargetPriority.cs
@@ -14,7 +14,7 @@ namespace SF3.Models.Structs.X1.Battle {
         [NameGetter(NamedValueType.Character)]
         [BulkCopy]
         public byte CharacterID {
-            get => (byte) Data.GetUInt8(_characterIdAddr);
+            get => Data.GetUInt8(_characterIdAddr);
             set => Data.SetUInt8(_characterIdAddr, value);
         }
     }
