+++ SF3Lib/Models/Structs/X1/Battle/TeamBitmask.cs
@@ -11,7 +11,7 @@ namespace SF3.Models.Structs.X1.Battle {
         [TableViewModelColumn(displayOrder: 0, displayFormat: "X4")]
         [BulkCopy]
         public ushort Bitmask {
-            get => (ushort) Data.GetUInt16(Address);
+            get => Data.GetUInt16(Address);
             set => Data.SetUInt16(Address, value);
         }
